Cleaning function file for tests to conform to the code quality
